--- src/main/java/nextstep/subway/applicaion/line/LineCreator.java
@@ -1,31 +1,0 @@
-package nextstep.subway.applicaion.line;
-
-import nextstep.subway.applicaion.dto.LineCreationRequest;
-import nextstep.subway.applicaion.dto.LineResponse;
-import nextstep.subway.domain.line.Line;
-import nextstep.subway.domain.line.LineRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class LineCreator {
-
-    private final LineRepository lineRepository;
-
-    public LineCreator(LineRepository lineRepository) {
-        this.lineRepository = lineRepository;
-    }
-
-    @Transactional
-    public LineResponse create(LineCreationRequest request) {
-        var line = new Line(
-                null,
-                request.getName(),
-                request.getColor(),
-                request.getUpStationId(),
-                request.getDownStationId(),
-                request.getDistance());
-
-        return LineResponse.fromLine(lineRepository.save(line));
-    }
-}
