--- src/main/java/nextstep/subway/applicaion/line/LineRemover.java
@@ -1,20 +1,0 @@
-package nextstep.subway.applicaion.line;
-
-import nextstep.subway.domain.line.LineRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class LineRemover {
-
-    private final LineRepository lineRepository;
-
-    public LineRemover(LineRepository lineRepository) {
-        this.lineRepository = lineRepository;
-    }
-
-    @Transactional
-    public void remove(Long lineId) {
-        lineRepository.deleteById(lineId);
-    }
-}
