+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -6,10 +6,8 @@ import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.LineCreationRequest;
 import nextstep.subway.applicaion.dto.LineModificationRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
-import nextstep.subway.applicaion.line.LineCreator;
-import nextstep.subway.applicaion.line.LineModifier;
 import nextstep.subway.applicaion.line.LineQueryService;
-import nextstep.subway.applicaion.line.LineRemover;
+import nextstep.subway.applicaion.line.LineService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,9 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class LineController {
 
-    private final LineCreator lineCreator;
-    private final LineModifier lineModifier;
-    private final LineRemover lineRemover;
+    private final LineService lineService;
     private final LineQueryService lineQueryService;
 
     @GetMapping("/lines")
@@ -41,7 +37,7 @@ public class LineController {
 
     @PostMapping("/lines")
     public ResponseEntity<LineResponse> createLine(@RequestBody LineCreationRequest request) {
-        var line = lineCreator.create(request);
+        var line = lineService.create(request);
         return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
     }
 
@@ -50,13 +46,13 @@ public class LineController {
             @PathVariable("lineId") Long lineId,
             @RequestBody LineModificationRequest request
     ) {
-        lineModifier.modify(lineId, request);
+        lineService.modify(lineId, request);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/lines/{lineId}")
     public ResponseEntity<Void> deleteLine(@PathVariable("lineId") Long lineId) {
-        lineRemover.remove(lineId);
+        lineService.remove(lineId);
         return ResponseEntity.noContent().build();
     }
 }
